Restructured includes to be in a subfolder and added `make install`.
