+++ Makefile.ossl
@@ -14,21 +14,25 @@
 
 # ---- QCBOR location ----
 # Adjust this to the location of QCBOR in your build environment
-QCBOR_INC= -I ../../QCBOR/master/inc
-QCBOR_LIB=../../QCBOR/master/libqcbor.a
+#QCBOR_INC= -I ../../QCBOR/master/inc
+#QCBOR_LIB=../../QCBOR/master/libqcbor.a
+QCBOR_INC= -I/usr/include -I/usr/local/include
+QCBOR_LIB= -l qcbor
 
 
 # ---- crypto configuration -----
 # Set up for OpenSSL. This may have to be adjusted for your build environment.
-CRYPTO_INC=-I ../../openssl/openssl-1.1.1b/include/openssl -I ../../openssl/openssl-1.1.1b/include
-CRYPTO_LIB=../../openssl/openssl-1.1.1b/libcrypto.a
+#CRYPTO_INC=-I ../../openssl/openssl-1.1.1b/include/openssl -I ../../openssl/openssl-1.1.1b/include
+#CRYPTO_LIB=../../openssl/openssl-1.1.1b/libcrypto.a
+CRYPTO_LIB= -l crypto
+CRYPTO_INC=
 CRYPTO_CONFIG_OPTS=-DT_COSE_USE_OPENSSL_CRYPTO
 CRYPTO_OBJ=crypto_adapters/t_cose_openssl_crypto.o
 CRYPTO_TEST_OBJ=test/t_cose_make_openssl_test_key.o
 
 
 # ---- compiler configuration -----
-C_OPTS=-Os -Wall -pedantic-errors -Wextra -Wshadow -Wparentheses -xc -std=c99
+C_OPTS=-Os -Wall -Wno-conversion -Werror -pedantic-errors -Wextra -Wshadow -Wparentheses -xc -std=c99
 
 
 # ---- T_COSE Config and test options ----
@@ -57,29 +61,41 @@ t_cose_test: main.o $(TEST_OBJ) libt_cose.a
 t_cose_basic_example_ossl:	examples/t_cose_basic_example_ossl.o libt_cose.a
 	cc -o $@ $^ $(QCBOR_LIB) $(CRYPTO_LIB)
 
+ifeq ($(PREFIX),)
+    PREFIX := /usr/local
+endif
+
+install: all
+	install -d $(DESTDIR)$(PREFIX)/lib/
+	install -m 644 libt_cose.a $(DESTDIR)$(PREFIX)/lib/
+	install -d $(DESTDIR)$(PREFIX)/include/t_cose
+	install -m 644 inc/t_cose/t_cose_common.h $(DESTDIR)$(PREFIX)/include/t_cose
+	install -m 644 inc/t_cose/q_useful_buf.h $(DESTDIR)$(PREFIX)/include/t_cose
+	install -m 644 inc/t_cose/t_cose_sign1_sign.h $(DESTDIR)$(PREFIX)/include/t_cose
+	install -m 644 inc/t_cose/t_cose_sign1_verify.h $(DESTDIR)$(PREFIX)/include/t_cose
+
 clean:
 	rm -f $(SRC_OBJ) $(TEST_OBJ) $(CRYPTO_OBJ) t_cose_basic_example_ossl t_cose_test libt_cose.a main.o
 
-
 # ---- public headers -----
-PUBLIC_INTERFACE=inc/t_cose_common.h inc/t_cose_sign1_sign.h inc/t_cose_sign1_verify.h
+PUBLIC_INTERFACE=inc/t_cose/t_cose_common.h inc/t_cose/t_cose_sign1_sign.h inc/t_cose/t_cose_sign1_verify.h
 
 # ---- source dependecies -----
-src/t_cose_util.o:	src/t_cose_util.h src/t_cose_standard_constants.h inc/t_cose_common.h src/t_cose_crypto.h
-src/t_cose_sign1_verify.o:	inc/t_cose_sign1_verify.h src/t_cose_crypto.h src/t_cose_util.h src/t_cose_parameters.h inc/t_cose_common.h src/t_cose_standard_constants.h
-src/t_cose_parameters.o: src/t_cose_parameters.h src/t_cose_standard_constants.h inc/t_cose_sign1_verify.h inc/t_cose_common.h
-src/t_cose_sign1_sign.o: inc/t_cose_sign1_sign.h src/t_cose_standard_constants.h src/t_cose_crypto.h src/t_cose_util.h inc/t_cose_common.h 
+src/t_cose_util.o:	src/t_cose_util.h src/t_cose_standard_constants.h inc/t_cose/t_cose_common.h src/t_cose_crypto.h
+src/t_cose_sign1_verify.o:	inc/t_cose/t_cose_sign1_verify.h src/t_cose_crypto.h src/t_cose_util.h src/t_cose_parameters.h inc/t_cose/t_cose_common.h src/t_cose_standard_constants.h
+src/t_cose_parameters.o: src/t_cose_parameters.h src/t_cose_standard_constants.h inc/t_cose/t_cose_sign1_verify.h inc/t_cose/t_cose_common.h
+src/t_cose_sign1_sign.o: inc/t_cose/t_cose_sign1_sign.h src/t_cose_standard_constants.h src/t_cose_crypto.h src/t_cose_util.h inc/t_cose/t_cose_common.h 
 
 
 # ---- test dependencies -----
 test/t_cose_test.o:	test/t_cose_test.h test/t_cose_make_test_messages.h src/t_cose_crypto.h $(PUBLIC_INTERFACE)
 test/t_cose_sign_verify_test.o:	test/t_cose_sign_verify_test.h test/t_cose_make_test_messages.h src/t_cose_crypto.h test/t_cose_make_test_pub_key.h $(PUBLIC_INTERFACE)
-test/t_cose_make_test_messages.o: test/t_cose_make_test_messages.h inc/t_cose_sign1_sign.h inc/t_cose_common.h src/t_cose_standard_constants.h src/t_cose_crypto.h src/t_cose_util.h
+test/t_cose_make_test_messages.o: test/t_cose_make_test_messages.h inc/t_cose/t_cose_sign1_sign.h inc/t_cose/t_cose_common.h src/t_cose_standard_constants.h src/t_cose_crypto.h src/t_cose_util.h
 test/run_test.o: test/run_test.h test/t_cose_test.h test/t_cose_hash_fail_test.h
-test/t_cose_make_openssl_test_key.o:	test/t_cose_make_test_pub_key.h src/t_cose_standard_constants.h inc/t_cose_common.h
+test/t_cose_make_openssl_test_key.o:	test/t_cose_make_test_pub_key.h src/t_cose_standard_constants.h inc/t_cose/t_cose_common.h
 
 # ---- crypto dependencies ----
-crypto_adapters/t_cose_openssl_crypto.o:	src/t_cose_crypto.h inc/t_cose_common.h src/t_cose_standard_constants.h inc/q_useful_buf.h
+crypto_adapters/t_cose_openssl_crypto.o:	src/t_cose_crypto.h inc/t_cose/t_cose_common.h src/t_cose_standard_constants.h inc/t_cose/q_useful_buf.h
 
 # ---- example dependencies ----
 t_cose_basic_example_ossl.o:	$(PUBLIC_INTERFACE)
