+++ src/t_cose_sign1_verify.c
@@ -9,10 +9,10 @@
  */
 
 
-#include "t_cose_sign1_verify.h"
-#include "qcbor.h"
+#include <qcbor/qcbor.h>
+#include <t_cose/t_cose_sign1_verify.h>
+#include <t_cose/q_useful_buf.h>
 #include "t_cose_crypto.h"
-#include "q_useful_buf.h"
 #include "t_cose_util.h"
 #include "t_cose_parameters.h"
