+++ test/t_cose_make_test_messages.c
@@ -9,7 +9,7 @@
  */
 
 #include "t_cose_make_test_messages.h"
-#include "qcbor.h"
+#include <qcbor/qcbor.h>
 #include "t_cose_crypto.h"
 #include "t_cose_util.h"
