+++ test/t_cose_make_test_messages.h
@@ -14,9 +14,9 @@
 
 #include <stdint.h>
 #include <stdbool.h>
-#include "qcbor.h"
-#include "t_cose_common.h"
-#include "t_cose_sign1_sign.h"
+#include <qcbor/qcbor.h>
+#include <t_cose/t_cose_common.h>
+#include <t_cose/t_cose_sign1_sign.h>
 
 
 #ifdef __cplusplus
